# Conflicts:
#	main.go
#	services/betService/betOrch.go

+++ models/user.go
@@ -4,6 +4,6 @@ type User struct {
 	ID        uint   `gorm:"primaryKey"`
 	DiscordID string `gorm:"uniqueIndex:user_guild_idx; size:64"`
 	GuildID   string `gorm:"uniqueIndex:user_guild_idx; size:64"`
-	Points    int
+	Points    float64
 	Username  *string
 }
